requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.2
pymongo>=4.6.0
celery>=5.3.4
redis>=5.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
playwright>=1.47.0
selenium>=4.23.1
webdriver-manager>=4.0.2
linkedin-scraper>=2.9.0
trafilatura>=1.6.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.2.27
langchain-openai>=0.0.2
langchain-google-genai>=2.0.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.1.0
httpx>=0.25.0 
